+++ Sources/FKCoreKit/Components/Extension/Foundation/TimeZone.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public extension TimeZone {
+  /// UTC time zone (zero offset from GMT).
+  static var fk_utc: TimeZone {
+    TimeZone(secondsFromGMT: 0)!
+  }
+}